tools/requirements.txt
Pillow
PySDL2
